requirements.txt
numpy==1.24.4
scipy==1.10.1
pandas==2.0.3
ultralytics==8.3.182
filterpy==1.4.5
networkx==3.1
typing_extensions==4.9.0
matplotlib==3.7.5
pillow==10.4.0
requests==2.32.4
psutil==5.9.8
PyYAML==6.0.2
tqdm==4.67.1
opencv-python==4.12.0.88
